+++ packages/foundation/decorators/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @praxisjs/decorators
 
+## 0.7.4
+
+### Patch Changes
+
+- Updated dependencies [994c581]
+  - @praxisjs/core@1.4.0
+
 ## 0.7.3
 
 ### Patch Changes